Full joystick axis control

+++ src/KeyBindings.h
@@ -49,7 +49,7 @@ namespace KeyBindings {
 
 		AxisBinding();
 		AxisBinding(Uint8 joystick, Uint8 axis, AxisDirection direction);
-		double GetValue();
+		float GetValue();
 		std::string Description() const;
 	};
